Fixed file naming issue. Created DiceRoller.java file for its class
